chore: Update PrTitleGenerateButton component to handle loading state
